requirements.txt
snowflake-snowpark-python
pyyaml
